+++ src/lib.rs
@@ -37,6 +37,9 @@ pub enum Error {
     #[error("Commands are only allowed at a tail position to avoid ambiguities")]
     NontailCommand(Ustr, ChicSpan),
 
+    #[error("Two adjacent terminals in a subword expression")]
+    SubwordSpaces(),
+
     #[error("UTF-8 conversion error")]
     FromUtf8Error(#[from] FromUtf8Error),